Steam games scraping
Querys for users that have one or more games in their wishlists, getting id list for inserts by id user, query for all games from users wishlist, ordered by stores, and an object for the results.
New function in ScrapingEJB that lets you add a cookie to a petition.
Some progress on the timer,