Bugfixes, fixed adding steam game to wishlist
Fixed a bug in AddGameOptions where if minimum price wasn't set, the program wouldn't reach the max price because of a NumberFormatException.
Fixed an error in AddGameOptions where an if statement was missing a !.
Fixed a bug in AddGameOptions, in case of a steam game, substring would get '/appid' instead of just the 'appid'.
Fixes in JspFunctions, so the prices can have up to 2 decimals, and inputs for max and min prices can now also have 2 decimals.
Changed the ResizeImgs javascript so now it resizes all of the images of all tables, not just the first one.
After logging in, SessionManagement takes you to the previous page now.