Changes in the database, updating game prices, bugfixing
Made some changes to the database, WISHLIST_GAME table is now directly related to the USER TABLE, removed the unnecessary LIST table, and changed the ID_LIST column from WISHLIST_GAME to ID_USER.
Changed all of the queries and functions that used ID_LIST to now use ID_USER.
Added a WishListGame2ScrapSteam object, that extends WishListGame2Scrap.
Renamed Game2Scrap url to storeUrl, so it is more clear what url it is.
ScrapedGame url is now separated into urlStore and urlGame.
Changed all of the functions that used links from ScrapedGame.
Added a query for getting games by user id.
Fixed some of the email formatting.
Fixed an error due to having spaces in the url of the game when searching for games in GOG.
Finished the getGame function in ScrapingSteam.