Price timeline saving
Added a table for game prices.
Added objects TimelineGame and TimelineGameDetailed.
Added mappers and querys for PRICE_TIMELINE table.
Added a trigger that if necessary deletes entries from PRICE_TIMELINE table.
Added a function in TimersEJB that adds and entry to the PRICE_TIMELINE table with current price and time.