Charts, and bugfixes
MyList now also sends a list of stores to the jsp.
MyList jsp now has a modal that has a list of games, games checkboxes can be marked, and then the petition can be send with a button, it will be used for generating graphs of choosen games.
Fixed a bug in TimersEJB where the ScrapedGame object wouldn't get WishListGame values if the price wasn't lower.
Changes in ClassPaths, now there are 2 private methods that generate the Strings.
Added bootstrap js.
Added a js for MyList, it is used for a check all checkboxes button.