JDK change, most of the chart generating logic
Changed the java jdk from 1.8 to 1.11 so that i can use some of the new functions for LocalDate objects.
Most of the logic of the PriceTimeline servlet.
Fixed a bug in Timeline.xml.
TimelineMapper methods now return ArrayLists, as they should have in the begging,
Added a new ejb for graph data handling.
Inserts into the PRICE_TIMELINE table will now be made once a day.
Changed public variables in ClassPaths that were used only inside the class to private.
Most of the PriceTimeline jsp logic.