Changes in db, finishing tests
Added a delete user query for tests.
Moved most of the columns from WISHLIST_GAME to WISHLIST_GAME_PRICE_TIMELINE and fixed all of the querys.
Added a catch for InternalServerErrorException in ScrapingSteam, it seems to happen randomly.
Added some constructors to some objects, so i can copy them easier.
Finished 90% of tests.