Revisió 4
Rename of WishlistGamePag to Pagination, as now it is used in various classes.
Some comments.
GameList, MyList and ReviewList pagination.
Switched to clasic for in loading steam games, as performance there is the most important.
Fixed and optimized a bit MailEJB.
Pagination is now iterable.
Added some more loggers to ScrapingInstantGaming, still trying to find the bug.
More fronted aditions, css file is a bit more organized.
Little optimizations in Checkboxes js, and also new functions for displaying errors if no checkboxes are selected.
Trying some stuff with session management.
Every site that had a form with checkboxes now has an implemented feature that shows an error if no checkboxes were checked.