Responsive price timeline, better updating page information
MyList and GameList servlets are now updating the list of games after any change to any of the items in the list, instead of everytime a correct parameter is send or the page gets refreshed.
Fixed getRevuewList query, it was missing a DISTINCT.
Removed some unnecessary imports.
PriceTimeline Graph is now fully responsive.