Added a site for viewing all of the current user reviews
As in the title, added a site for viewing all of the current user reviews, so it's easier to manage all of the reviews, there are options for updating and deleting a review.
Fixed a bug in GameInfo.
Added a message for cookies, and some javascript for deciding whether the message should be shown or not.
bootstrap.min.js is now included in all of the sites.