Fixed timeline graph names not matching lines
getTimelineByUrlDetailed now returns name of the game with sotre name.
"names" ArrayList is no longer being used in PriceTimeline class, now HashTable key is used for saving game, and store name.
Fixed a bug in GraphDatesEJB, where if it was the last day of the month, the month wouldnt switch to the next one and reset day to 1, instead it would leave the day on for exemple 32.