Bugfixing, sending email troubleshooting
Fixed a bug in GameInfo, RequestDispatcher in catch was redirecting to the wrong class.
UserReviews now can redirect to two different pages, GameInfo and UserReviews, depending on the referer.
Some small optimizations.
Added a function in MyLogger for loggin the whole Exception, it's not efficient, but it was necessary for troubleshooting email problems.